+++ static/app/views/detectors/list/common/detectorListHeader.tsx
